+++ django_geoip/management/commands/geoip_update.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from django.core.management.base import BaseCommand
 from optparse import make_option
+import logging
 from django_geoip.management.ipgeobase import IpGeobase
 
 
@@ -14,9 +15,21 @@ class Command(BaseCommand):
         ),
     )
 
+    def get_logger(self, verbosity):
+        logger = logging.getLogger('import')
+        logger.addHandler(logging.StreamHandler())
+        VERBOSITY_MAPPING = {
+            0: logging.CRITICAL, # no
+            1: logging.INFO, # means normal output (default)
+            2: logging.DEBUG, # means verbose output
+            3: logging.DEBUG, # means very verbose output
+        }
+        logger.setLevel(VERBOSITY_MAPPING[int(verbosity)])
+        return logger
+
     def handle(self, *args, **options):
-        print u'Updating geoip data...'
-        backend = IpGeobase()
+        logger = self.get_logger(options['verbosity'])
+        backend = IpGeobase(logger=logger)
 
         if options.get('clear'):
             backend.clear_database()