+++ runtests.py
@@ -5,4 +5,5 @@ if __name__ == "__main__":
 
     from django.core.management import execute_from_command_line
 
-    execute_from_command_line(['runtests.py', 'test'])
+    argv = ['runtests.py', 'test'] + sys.argv[1:]
+    execute_from_command_line(argv)